[Serve] Deflake test_deployment_actors: isolate controller-restart tests
Move test_deployment_actor_survives_controller_restart and
test_controller_restart_preserves_mutated_actor_state from
test_deployment_actors.py to test_gcs_failure.py.

Killing the Serve controller with ray.kill(controller, no_restart=False)
can destabilize the Ray cluster under CI resource constraints. Buildkite
CI logs show the next test (test_deployment_actor_restarts_on_crash)
inherits the broken session-scoped cluster and hangs for 831s — a single
test consuming nearly the entire 900s Bazel timeout. Isolating these
tests in a separate file (with a dedicated function-scoped fixture)
prevents cascading failures.

Investigation: https://gist.github.com/harshit-anyscale/878b21040ffe3e4287a4f135efa9cb8d
