+++ python/ray/serve/tests/test_gcs_failure.py
@@ -385,5 +385,193 @@ def test_proxy_router_updated_replicas_then_gcs_failure(serve_ha):
     assert len(returned_pids) == 2
 
 
+# ---------------------------------------------------------------------------
+# Controller-restart tests for deployment-scoped actors.
+#
+# These live here (not in test_deployment_actors.py) because killing the Serve
+# controller with ``ray.kill(controller, no_restart=False)`` can destabilize
+# the Ray cluster under CI resource constraints — subsequent tests that share
+# the same session-scoped ``serve_instance`` fixture may see GCS connectivity
+# errors and hang for minutes.  Isolating them in a separate file prevents
+# cascading failures.
+#
+# Uses a dedicated function-scoped fixture (not the session-scoped
+# ``serve_instance`` from conftest) to avoid conflicts with the
+# function-scoped ``serve_ha`` fixture used by the GCS-failure tests above.
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture(scope="function")
+def serve_instance_for_controller_restart():
+    """Function-scoped Serve instance for controller-restart tests.
+
+    Avoids conflicts with the function-scoped ``serve_ha`` fixture
+    (which independently calls ``ray.init``/``ray.shutdown``).
+    """
+    ray.init(
+        num_cpus=36,
+        namespace="default_test_namespace",
+        _metrics_export_port=9999,
+        _system_config={"metrics_report_interval_ms": 1000, "task_retry_delay_ms": 50},
+    )
+    serve.start(proxy_location="HeadOnly", http_options={"host": "0.0.0.0"})
+    yield _get_global_client()
+    serve.shutdown()
+    ray.shutdown()
+
+
+@ray.remote
+class _ControllerRestartSharedCounter:
+    """Trivial counter used by controller-restart tests below."""
+
+    def __init__(self, start: int = 0):
+        self.count = start
+
+    def increment(self):
+        self.count += 1
+        return self.count
+
+    def get(self):
+        return self.count
+
+    def ray_actor_id(self) -> str:
+        return ray.get_runtime_context().get_actor_id()
+
+
+def _get_deployment_actor_names_for_app(app_name: str, deployment_name: str) -> list:
+    """Return deployment actor names for a specific app and deployment."""
+    from ray.serve._private.constants import SERVE_DEPLOYMENT_ACTOR_PREFIX
+    from ray.util.state import list_actors
+
+    prefix = f"{SERVE_DEPLOYMENT_ACTOR_PREFIX}{app_name}::{deployment_name}::"
+    return [
+        a["name"]
+        for a in list_actors(filters=[("state", "=", "ALIVE")])
+        if a.get("name", "").startswith(prefix)
+    ]
+
+
+def test_deployment_actor_survives_controller_restart(
+    serve_instance_for_controller_restart,
+):
+    """Deployment actors are detached and survive controller restart.
+
+    After controller recovers from checkpoint, it discovers existing
+    deployment actors (ActorAlreadyExistsError) and marks them ready.
+    App continues to work with the same Ray actor process (matching ``ray_actor_id``),
+    not a recreation.
+    """
+    from ray.serve._private.test_utils import get_application_url, request_with_retries
+
+    @serve.deployment(
+        deployment_actors=[
+            DeploymentActorConfig(
+                name="counter",
+                actor_class=_ControllerRestartSharedCounter,
+                init_kwargs={"start": 42},
+            ),
+        ],
+    )
+    class MyDeployment:
+        def __call__(self):
+            counter = serve.get_deployment_actor("counter")
+            val = ray.get(counter.get.remote())
+            aid = ray.get(counter.ray_actor_id.remote())
+            return f"{val},{aid}"
+
+    serve.run(
+        MyDeployment.bind(),
+        name="app",
+        route_prefix="/survives_controller_restart_da",
+    )
+
+    def _parse_val_actor_id(text: str) -> tuple[str, str]:
+        val, aid = text.split(",", 1)
+        return val, aid
+
+    actor_id_before = None
+    for _ in range(5):
+        resp = request_with_retries(timeout=30, app_name="app")
+        val, aid = _parse_val_actor_id(resp.text)
+        assert val == "42"
+        if actor_id_before is None:
+            actor_id_before = aid
+        else:
+            assert aid == actor_id_before
+
+    actor_names_before = _get_deployment_actor_names_for_app("app", "MyDeployment")
+    assert len(actor_names_before) == 1
+
+    ray.kill(serve_instance_for_controller_restart._controller, no_restart=False)
+
+    wait_for_condition(
+        lambda: get_application_url("HTTP", "app", use_localhost=True) is not None
+    )
+    for _ in range(10):
+        resp = request_with_retries(timeout=30, app_name="app")
+        val, aid = _parse_val_actor_id(resp.text)
+        assert val == "42"
+        assert aid == actor_id_before
+
+    actor_names_after = _get_deployment_actor_names_for_app("app", "MyDeployment")
+    assert actor_names_after == actor_names_before
+
+
+def test_controller_restart_preserves_mutated_actor_state(
+    serve_instance_for_controller_restart,
+):
+    """Controller restart preserves mutated deployment actor state.
+
+    Strengthens test_deployment_actor_survives_controller_restart: that test
+    only checks the init value (start=42) survives. This test mutates the
+    actor's state (increments counter beyond init value) before the restart
+    and verifies the mutations survive — proving the *same* actor instance
+    continues running, not a freshly created one.
+    """
+    from ray.serve._private.constants import SERVE_NAMESPACE
+    from ray.serve._private.test_utils import get_application_url, request_with_retries
+
+    @serve.deployment(
+        deployment_actors=[
+            DeploymentActorConfig(
+                name="counter",
+                actor_class=_ControllerRestartSharedCounter,
+                init_kwargs={"start": 0},
+            ),
+        ],
+    )
+    class MyDeployment:
+        def __call__(self):
+            counter = serve.get_deployment_actor("counter")
+            return str(ray.get(counter.get.remote()))
+
+    serve.run(
+        MyDeployment.bind(),
+        name="app",
+        route_prefix="/preserves_mutated_state_da",
+    )
+    resp = request_with_retries(timeout=30, app_name="app")
+    assert resp.text == "0"
+
+    # Mutate actor state directly (bypass replica to isolate the test)
+    actor_names = _get_deployment_actor_names_for_app("app", "MyDeployment")
+    assert len(actor_names) == 1
+    handle = ray.get_actor(actor_names[0], namespace=SERVE_NAMESPACE)
+    for _ in range(5):
+        ray.get(handle.increment.remote())
+    assert ray.get(handle.get.remote()) == 5
+
+    ray.kill(serve_instance_for_controller_restart._controller, no_restart=False)
+
+    wait_for_condition(
+        lambda: get_application_url("HTTP", "app", use_localhost=True) is not None
+    )
+
+    # Mutated state (5) must survive, not reset to init value (0)
+    for _ in range(5):
+        resp = request_with_retries(timeout=30, app_name="app")
+        assert resp.text == "5"
+
+
 if __name__ == "__main__":
     sys.exit(pytest.main(["-v", "-s", __file__]))
